+++ src/cascader/cascader-search-option.component.ts
@@ -1,21 +1,9 @@
 import { NgClass, NgTemplateOutlet } from '@angular/common';
-import {
-    ChangeDetectionStrategy,
-    Component,
-    EventEmitter,
-    HostBinding,
-    HostListener,
-    Input,
-    OnInit,
-    Output,
-    TemplateRef,
-    ViewEncapsulation
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, HostListener, OnInit, TemplateRef, ViewEncapsulation, input, output } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ThyBreadcrumb, ThyBreadcrumbItem } from 'ngx-tethys/breadcrumb';
 import { ThyCheckbox } from 'ngx-tethys/checkbox';
 import { ThyFlexibleText } from 'ngx-tethys/flexible-text';
-import { ThyIcon } from 'ngx-tethys/icon';
 import { ThyCascaderSearchOption } from './types';
 import { SafeAny } from 'ngx-tethys/types';
 import { coerceBooleanProperty } from 'ngx-tethys/util';
@@ -29,35 +17,33 @@ import { coerceBooleanProperty } from 'ngx-tethys/util';
     // eslint-disable-next-line @angular-eslint/component-selector
     selector: '[thy-cascader-search-option]',
     templateUrl: './cascader-search-option.component.html',
-    imports: [NgClass, ThyFlexibleText, ThyCheckbox, ThyBreadcrumb, ThyBreadcrumbItem, FormsModule, NgTemplateOutlet]
+    imports: [NgClass, ThyFlexibleText, ThyCheckbox, ThyBreadcrumb, ThyBreadcrumbItem, FormsModule, NgTemplateOutlet],
+    host: {
+        class: 'thy-cascader-search-list-item px-4 d-flex align-items-center cursor-pointer',
+        '[class.multiple]': 'multiple()',
+        '[class.active]': 'active()'
+    }
 })
 export class ThyCascaderSearchOptionComponent implements OnInit {
-    @Input() option: ThyCascaderSearchOption;
+    readonly option = input<ThyCascaderSearchOption>();
 
-    @HostBinding('class.multiple')
-    @Input({ transform: coerceBooleanProperty })
-    multiple = false;
+    readonly multiple = input(false, { transform: coerceBooleanProperty });
 
-    @Input({ transform: coerceBooleanProperty })
-    isOnlySelectLeaf = true;
-    @HostBinding('class') className = 'thy-cascader-search-list-item px-4 d-flex align-items-center cursor-pointer';
+    readonly isOnlySelectLeaf = input(true, { transform: coerceBooleanProperty });
 
-    @HostBinding('class.active')
-    @Input({ transform: coerceBooleanProperty })
-    active: boolean = false;
+    readonly active = input(false, { transform: coerceBooleanProperty });
 
-    @Input()
-    optionRender: TemplateRef<SafeAny>;
+    readonly optionRender = input<TemplateRef<SafeAny>>(undefined);
 
     @HostListener('click', ['$event'])
     public toggleClick($event: Event) {
-        if (this.multiple) {
+        if (this.multiple()) {
             return;
         }
-        this.toggleSelectChange.emit(this.option);
+        this.toggleSelectChange.emit(this.option());
     }
 
-    @Output() toggleSelectChange: EventEmitter<ThyCascaderSearchOption> = new EventEmitter();
+    readonly toggleSelectChange = output<ThyCascaderSearchOption>();
 
     constructor() {}
 
@@ -65,13 +51,13 @@ export class ThyCascaderSearchOptionComponent implements OnInit {
 
     public clickCheckbox(event: Event) {
         // 已选中的在搜索情况下不能取消选择
-        if (this.active) {
+        if (this.active()) {
             event.preventDefault();
             event.stopPropagation();
         }
     }
 
     public toggleOption(value: boolean) {
-        this.toggleSelectChange.emit(this.option);
+        this.toggleSelectChange.emit(this.option());
     }
 }